Users/friends/invitations
Actualize methods for pagination

+++ src/main/java/com/example/persimmoncocktails/services/FriendsService.java
@@ -2,6 +2,7 @@ package com.example.persimmoncocktails.services;
 
 import com.example.persimmoncocktails.dao.FriendsDao;
 import com.example.persimmoncocktails.dao.FriendshipInvitationDao;
+import com.example.persimmoncocktails.dao.PersonDao;
 import com.example.persimmoncocktails.dtos.friend.FriendResponseDto;
 import com.example.persimmoncocktails.exceptions.IncorrectNameFormat;
 import com.example.persimmoncocktails.exceptions.IncorrectRangeNumberFormat;
@@ -16,6 +17,7 @@ import java.util.List;
 @AllArgsConstructor
 public class FriendsService {
     FriendsDao friendsDao;
+    PersonDao personDao;
     FriendshipInvitationDao friendshipInvitationDao;
 
     public List<FriendResponseDto> searchPersonsByNameSubstring(String substring, Long pageNumber) { // return list of persons by page, who is not an administrator and moderator
@@ -26,22 +28,27 @@ public class FriendsService {
 
     public List<FriendResponseDto> searchPersonsByNameSubstringWithoutFriends(Long personId, String substring, Long pageNumber) {
         // return list of persons by page, who is not an administrator and moderator and is not a friend
+        if (!personDao.existsById(personId)) throw new NotFoundException("Person");
         if (!AuthorizationService.nameIsValid(substring)) throw new IncorrectNameFormat();
         if (pageNumber < 0) throw new IncorrectRangeNumberFormat("of page");
         return friendsDao.searchPersonsByNameSubstringWithoutFriends(personId, "%" + substring + "%", pageNumber);
     }
 
     public void removeFriendById(Long personIdInitiator, Long friendId) {
+        if (!personDao.existsById(friendId)) throw new NotFoundException("Person");
         friendsDao.removeFriendById(personIdInitiator, friendId);
     }
 
     public List<FriendResponseDto> getPersonFriends(Long personId, Long pageNumber) {
+        if (!personDao.existsById(personId)) throw new NotFoundException("Person");
         if (pageNumber < 0) throw new IncorrectRangeNumberFormat("of page");
-        return friendsDao.getPersonFriends(personId, pageNumber);
+        return friendsDao.getListFriendByNameSubstring(personId, "%%", pageNumber);
     }
 
     public List<FriendResponseDto> getListFriendsBySubstring(Long personId, String substring, Long pageNumber) {
+        if (!personDao.existsById(personId)) throw new NotFoundException("Person");
         if (pageNumber < 0) throw new IncorrectRangeNumberFormat("of page");
+        if (substring.equals("")) return getPersonFriends(personId, pageNumber); // return list of friends without search
         if (!AuthorizationService.nameIsValid(substring)) throw new IncorrectNameFormat();
         return friendsDao.getListFriendByNameSubstring(personId, "%" + substring + "%", pageNumber);
     }
@@ -49,9 +56,26 @@ public class FriendsService {
     public void addFriend(Long personIdReceiver, Long personIdInitiator) {
         if (personIdInitiator.equals(personIdReceiver))
             throw new WrongCredentialsException("The user cannot have friendship with himself.");
-        if (!friendshipInvitationDao.friendshipHasInInvitation(personIdReceiver, personIdInitiator))
+        if (!friendshipInvitationDao.friendshipInvitationPairExists(personIdReceiver, personIdInitiator))
             throw new NotFoundException("friendship invitation");
         friendsDao.addFriend(personIdInitiator, personIdReceiver);
         friendshipInvitationDao.deleteFriendshipInvitation(personIdReceiver, personIdInitiator);
+    }
+
+    public Boolean isPersonsBySubstringWithoutFriendsExists(Long personId, String substring, Long pageNumber) {
+        if (!personDao.existsById(personId)) throw new NotFoundException("Person");
+        if (!AuthorizationService.nameIsValid(substring)) throw new IncorrectNameFormat();
+        if (pageNumber < 0) throw new IncorrectRangeNumberFormat("of page");
+        return friendsDao.isPersonsBySubstringWithoutFriendsExists(personId, "%" + substring + "%", pageNumber);
+    }
+
+    public Long numberOfPagesPersonsBySubstringWithoutFriends(Long personId, String substring) {
+        if (!personDao.existsById(personId)) throw new NotFoundException("Person");
+        return friendsDao.getNumberOfPagePersonsBySubstringWithoutFriends(personId, "%"+substring+"%");
+    }
+
+    public Long numberOfPagesFriendsBySubstring(Long personId, String substring) {
+        if (!personDao.existsById(personId)) throw new NotFoundException("Person");
+        return friendsDao.getNumberOfPageFriends(personId, "%"+substring+"%");
     }
 }
